requirements.txt
# --- 核心框架 ---
langchain
langchain-community
langchain-core
langchain-text-splitters

# --- 智能体编排 (LangGraph) ---
langgraph
langgraph-checkpoint
langgraph-prebuilt

# --- LLM 模型接口 (新增/修改) ---
langchain-openai
python-dotenv

# --- 向量数据库与检索 ---
langchain-qdrant
qdrant-client
fastembed

# --- 嵌入模型 (本地运行) ---
langchain-huggingface
sentence-transformers
huggingface-hub

# --- 文档处理 ---
pymupdf4llm
PyMuPDF

# --- 用户界面 ---
gradio

# --- 基础工具 ---
pydantic
httpx